+++ app/src/main/AndroidManifest.xml
@@ -34,67 +34,67 @@
         <receiver android:name=".NotificationReceiver" />
 
         <activity
-            android:name=".activitys.TelaPesquisa"
+            android:name=".activities.TelaPesquisa"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaScanner"
+            android:name=".activities.TelaScanner"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaInfoGuia"
+            android:name=".activities.TelaInfoGuia"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaGuias"
+            android:name=".activities.TelaGuias"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaInfoObra"
+            android:name=".activities.TelaInfoObra"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaInfoArtista"
+            android:name=".activities.TelaInfoArtista"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaInfoNoticia"
+            android:name=".activities.TelaInfoNoticia"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaAreaRestrita"
+            android:name=".activities.TelaAreaRestrita"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaInfoMuseu"
+            android:name=".activities.TelaInfoMuseu"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaEditarPerfil"
+            android:name=".activities.TelaEditarPerfil"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaInfoGenero"
+            android:name=".activities.TelaInfoGenero"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaPrincipal"
+            android:name=".activities.TelaPrincipal"
             android:exported="false" />
         <activity
-            android:name=".activitys.TelaBemVindo"
+            android:name=".activities.TelaBemVindo"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".activitys.TelaCadastroFotoPerfil"
+            android:name=".activities.TelaCadastroFotoPerfil"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".activitys.TelaCadastroInteresses"
+            android:name=".activities.TelaCadastroInteresses"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".activitys.TelaCadastro2"
+            android:name=".activities.TelaCadastro2"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".activitys.TelaCadastro"
+            android:name=".activities.TelaCadastro"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".activitys.TelaLogin"
+            android:name=".activities.TelaLogin"
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".activitys.TelaSplashScreen"
+            android:name=".activities.TelaSplashScreen"
             android:exported="true"
             android:screenOrientation="portrait">
             <intent-filter>
