+++ app/src/main/java/com/aula/leontis/adapters/AdapterObra.java
@@ -6,14 +6,13 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
-import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.aula.leontis.Geral;
 import com.aula.leontis.R;
-import com.aula.leontis.activitys.TelaInfoObra;
+import com.aula.leontis.activities.TelaInfoObra;
 import com.aula.leontis.models.obra.Obra;
 import com.bumptech.glide.Glide;
