+++ app/src/main/res/layout-sw600dp/tela_bem_vindo.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/bege_claro"
-    tools:context=".activitys.TelaBemVindo">
+    tools:context=".activities.TelaBemVindo">
 
     <TextView
         android:id="@+id/textView10"
