+++ app/src/main/res/layout/tela_cadastro.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/bege_claro"
-    tools:context=".activitys.TelaCadastro">
+    tools:context=".activities.TelaCadastro">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
