+++ app/src/main/res/layout/tela_cadastro2.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/bege_claro"
-    tools:context=".activitys.TelaCadastro2">
+    tools:context=".activities.TelaCadastro2">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
