+++ app/src/main/res/layout/tela_cadastro_foto_perfil.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/bege_claro"
-    tools:context=".activitys.TelaCadastroFotoPerfil">
+    tools:context=".activities.TelaCadastroFotoPerfil">
 
     <TextView
         android:id="@+id/textView8"
