+++ app/src/main/res/layout/tela_cadastro_interesses.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/bege_claro"
-    tools:context=".activitys.TelaCadastroInteresses">
+    tools:context=".activities.TelaCadastroInteresses">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_generos"
