+++ app/src/main/res/layout/tela_editar_perfil.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activitys.TelaEditarPerfil">
+    tools:context=".activities.TelaEditarPerfil">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
