+++ app/src/main/res/layout/tela_guias.xml
@@ -7,7 +7,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activitys.TelaGuias">
+        tools:context=".activities.TelaGuias">
 
         <androidx.cardview.widget.CardView
             android:id="@+id/cardView"
