+++ app/src/main/res/layout/tela_info_artista.xml
@@ -7,7 +7,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activitys.TelaInfoMuseu">
+        tools:context=".activities.TelaInfoMuseu">
 
         <ImageView
             android:id="@+id/fotoArtista"
