+++ app/src/main/res/layout/tela_info_genero.xml
@@ -7,7 +7,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activitys.TelaInfoGenero">
+    tools:context=".activities.TelaInfoGenero">
 
     <ImageView
         android:id="@+id/fotoGenero"
