+++ app/src/main/res/layout/tela_info_guia.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activitys.TelaInfoGuia">
+    tools:context=".activities.TelaInfoGuia">
 
     <ImageButton
         android:id="@+id/btnFiltrar"
