+++ app/src/main/res/layout/tela_info_museu.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activitys.TelaInfoMuseu">
+    tools:context=".activities.TelaInfoMuseu">
 
     <ScrollView
         android:layout_width="match_parent"
