+++ app/src/main/res/layout/tela_info_noticia.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activitys.TelaAreaRestrita">
+    tools:context=".activities.TelaAreaRestrita">
 
     <ImageButton
         android:id="@+id/btn_voltar_noticia"
