+++ app/src/main/res/layout/tela_info_obra.xml
@@ -13,7 +13,7 @@
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            tools:context=".activitys.TelaInfoObra">
+            tools:context=".activities.TelaInfoObra">
 
             <ImageView
                 android:id="@+id/fotoObra"
