+++ app/src/main/res/layout/tela_pesquisa.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activitys.TelaPesquisa">
+    tools:context=".activities.TelaPesquisa">
     <ImageButton
         android:id="@+id/btnVoltar"
         android:layout_width="wrap_content"
