+++ app/src/main/res/layout/tela_principal.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activitys.TelaPrincipal">
+    tools:context=".activities.TelaPrincipal">
 
     <androidx.fragment.app.FragmentContainerView
         android:layout_width="match_parent"
