+++ app/src/main/res/layout/tela_scanner.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activitys.TelaScanner">
+    tools:context=".activities.TelaScanner">
 
     <androidx.camera.view.PreviewView
         android:id="@+id/viewFinder"
